+++ src/ops/skip_last.rs
@@ -122,11 +122,11 @@ mod test {
   }
 
   #[test]
-  fn into_shared() {
+  fn ininto_shared() {
     observable::from_iter(0..100)
       .skip_last(5)
       .skip_last(5)
-      .to_shared()
+      .into_shared()
       .subscribe(|_| {});
   }
   #[bench]